requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
DateTime==5.1
fonttools==4.40.0
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
pyparsing==3.0.9
pytest==7.3.2
pytest-warnings==0.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
zope.interface==6.0
pytz==2023.3
pathlib==1.0.1
